Issue with catch up compensation and leave type for counters
Good job for this app but i've got an issue if i keep this query.This query give me a sum of durations of the leave type 0, so it should be a const declared in the config file i presume. For my example, type 0 is paid leave so i have wrong counter in the result array. it's a temporary solution which work for me because i don't use overtime. Thank you.